Move ConfigProviderTest class to header, so we can include it in multiple tests, add removed test to test ref device module, and some refactoring
