+++ core/opendaq/opendaq/tests/CMakeLists.txt
@@ -2,6 +2,7 @@ set(MODULE_NAME ${SDK_TARGET_NAME})
 set(TEST_APP test_${MODULE_NAME})
 
 set(TEST_HEADERS test_helpers.h
+                 test_config_provider.h
 )
 
 set(TEST_SOURCES test_factories.cpp
